Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. Although the name of the repository does match the sketch name, when GitHub's popular Clone or download > Download ZIP feature is used to download the contents of a repository the branch/tag/commit name is appended to the folder name, causing a mismatch.

When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "G27_Pedals_and_Shifter.ino" needs to be inside a sketch folder named "G27_Pedals_and_Shifter".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file G27_Pedals_and_Shifter.ino to the new folder, leaving behind the other source files. This causes compilation of the sketch to fail:

fatal error: ./lib/G27PedalsShifter.h: No such file or directory